Fix style
