+++ example/.env.example
@@ -1,0 +1,3 @@
+REACT_APP_GRAPH_API_URL=http://localhost:8088/api
+REACT_APP_GRAPH_API_NODES_PATH=nodes
+REACT_APP_GRAPH_API_RELATIONSHIPS_PATH=relationships
