+++ example/src/demos/Minimalist.tsx
@@ -1,13 +1,14 @@
 import React from "react";
 import ReactGraph from "react-graph";
 import { PageWrapper } from "../components/common/Organisms";
+import { root, nodes, relationships } from "../assets/mocks/random";
 
-const Minimalist = (props: any) => (
+const Minimalist = () => (
   <PageWrapper>
     <ReactGraph
-      initialState={props.root}
-      nodes={props.nodes}
-      relationships={props.relationships}
+      initialState={root}
+      nodes={nodes}
+      relationships={relationships}
       width="100%"
       height="100%"
       hasLegends
