+++ example/src/demos/index.tsx
@@ -1,4 +1,4 @@
-import { Dashboard } from "./Dashboard";
+import { DashboardContainer as Dashboard } from "./dashboard/DashboardContainer";
 import { Minimalist } from "./Minimalist";
 
 export { Dashboard, Minimalist };
